requirements.txt
Keras==2.1.4
numpy==1.11.0
opencv_python==3.4.1.15
matplotlib==2.0.2
pandas==0.21.0
Pillow==5.3.0
scikit_learn==0.20.0
